# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Image Studio (Free Gemini Web Version)",
    "description": "Generate and edit images via your personal Google account (Gemini Web).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://YOUR_NGROK_URL_HERE",
      "description": "Your local Ngrok or Cloudflare tunnel"
    }
  ],
  "paths": {
    "/configure_google_login": {
      "post": {
        "operationId": "configureGoogleLogin",
        "summary": "Set your personal Google Account cookie to authenticate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "secure1psid": {
                    "type": "string",
                    "description": "The value of the __Secure-1PSID cookie from gemini.google.com"
                  }
                },
                "required": [
                  "secure1psid"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/generate_gemini_image": {
      "post": {
        "operationId": "generateGeminiImage",
        "summary": "Generate a new image using Google Gemini",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Description of the image to generate"
                  },
                  "model": {
                    "type": "string"
                  },
                  "quality": {
                    "type": "string",
                    "enum": [
                      "high",
                      "fast"
                    ]
                  }
                },
                "required": [
                  "prompt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/edit_image": {
      "post": {
        "operationId": "editGeminiImage",
        "summary": "Edit an existing image based on text instructions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "What to change or edit in the image"
                  },
                  "imagePath": {
                    "type": "string",
                    "description": "Absolute path to the local image file to edit"
                  }
                },
                "required": [
                  "prompt",
                  "imagePath"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/status": {
      "get": {
        "operationId": "getStatus",
        "summary": "Check server configuration status",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/configure_storage": {
      "post": {
        "summary": "Configure local storage directories for images and videos",
        "operationId": "configureStorage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageDir": {
                    "type": "string"
                  },
                  "videoDir": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Storage successfully configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}